Financial Agent/requirements.txt
phidata
python-dotenv
yfinance
duckduckgo-search
packaging
fastapi
uvicorn
groq